Adding spotless plugin
 -- removing app module by moving src to root. This allows spotless to fetch google code format plugin correctly
