+++ docker_interop/settings.gradle
@@ -8,4 +8,3 @@
  */
 
 rootProject.name = 'docker_interop'
-include('app')
