+++ docker_interop/src/main/resources/.env
