+++ docker_interop/src/main/resources/besu_genesis_template.json
