+++ docker_interop/src/main/resources/docker-compose-template.yml
