+++ docker_interop/src/main/resources/quorum_genesis_template.json
