+++ docker_interop/src/main/resources/run_geth.sh
